Bring more sanity to integration tests.
Factor out utils to a separate lib.
